Allow querying of none FN terms (i.e. TEL).
Juergen Edner asked if it was possible to query against non full
name terms, such as a telephone number. This allows it (if your
carddav supports it -- davical is broken and doesn't).

Added a query argument to the command line.

Passed the query off to the http request/report. Reformatted the
http request (added more whitespace).

Initialised variables to 0 and general clean-ups.