Remove generated files
Since the application cannot be built from the repository directly
without using autotools, there is no point in partially including files
that the autotools can regenerate, making source control harder.

This patch removes build-aux/ and src/config.h.in, adjusts .gitignore
accordingly and adds instructions to the README about how to
rebuild from scratch and a pointer to use the release tarball if you
don't want to or can't use the autotools.