Ability to run `-V` and `-h` before `decrypt()`.
We used to read the default rc file then parse the command line,
and then read a user supplied config file.

We changed this logic to parse the command line first so as to
be able to use `-V` and `-h` and exit quickly. Then to read a
configuration file. If one is not supplied as an argument, we
just read the default one (~/.mcdsrc). The only option/argument
that is specified in both the rc file and arguments is the URL,
and I originally made a check to make sure to not overwrite it
if it was an argument when reading the rc file.