Unfold vCard before using it.
Follow the RFC by unfolding folded vCard lines (CRLF WSP) before using
the vCard. This is done in place as we will be accessing all the data
immediately anyway as we pass the automaton over it so it is likely to
stay in cache.

doxygen tweak to new xregcomp function

Treat CR as optional when unfolding vCard

This is a pipelined approach - doing the unfolding as a separate
step rather than including special handling of continuation lines in
the later handling, for clarity.