Unfold vCard before using it.
Follow the RFC by unfolding folded vCard lines (CRLF WSP) before using
the vCard. This is done in place as we will be accessing all the data
immediately anyway as we pass the automata over it so it is likely to
stay in cache.

This pipelined approach seems easier than special handling of
continuation lines and follows the spirit of the specification.